+++ scripts/targets/batch/html5.yaml
@@ -21,7 +21,7 @@ inherits: "../../templates/html5.yaml"
 batch: 
     - caption: "Running on batch mode for ~{$CURRENT_FOLDER_NAME}:"       
     - local:                             
-        - path: "{$APP_FOLDER_PATH}/../test/samples/private/html5"
+        - path: "<SOME_PATH>"
 
     - copy_detector:            
         type: "HTML"
