+++ scripts/targets/batch/permissions.yaml
@@ -26,7 +26,7 @@ batch:
     - concurrent: 1 #cannot run in parallel due permissions over users
     
     - local:                             
-        - path: "{$APP_FOLDER_PATH}/../test/samples/private/permissions"
+        - path: "<SOME_PATH>"
         - vars:
             current_host: "localhost"
             current_user: "postgres"
