+++ scripts/targets/batch/xml_validation.yaml
@@ -21,7 +21,7 @@ inherits: "../../templates/xml_validation.yaml"
 batch: 
     - caption: "Running on batch mode for ~{$CURRENT_FOLDER_NAME}:"       
     - local:                             
-        - path: "{$APP_FOLDER_PATH}/../test/samples/private/xml"
+        - path: "<SOME_PATH>"
 
     - copy_detector:            
         type: "XML"
