+++ scripts/targets/single/html5.yaml
@@ -19,4 +19,4 @@
 inherits: "../../templates/html5.yaml"  
 single:
   local:
-    folder:  "{$APP_FOLDER_PATH}/../test/samples/private/html5/Student Name 1"
+    folder:  "<SOME_FOLDER>"
