+++ scripts/targets/single/web_syndication.yaml
@@ -19,4 +19,4 @@
 inherits: "../../templates/web_syndication.yaml"  
 single:
   local:
-    folder:  "{$APP_FOLDER_PATH}/../test/samples/private/web_syndication/Student Name 1"
+    folder:  "<SOME_FOLDER>"
