+++ test/main/Script.cs/Real/Mineseeker.cs
@@ -23,12 +23,12 @@
 using System.Collections.Generic;
 using NUnit.Framework;
 
-namespace AutoCheck.Test
+namespace AutoCheck.Test.Real
 {    
     [Parallelizable(ParallelScope.All)]    
-    public class RealMineseeker : Test
+    public class Mineseeker : Test
     { 
-        public RealMineseeker(): base("script/real"){
+        public Mineseeker(): base("script/real"){
         }      
 
         [Test, Category("Mineseeker"), Category("Real"), Category("Local")]    
@@ -51,7 +51,8 @@ namespace AutoCheck.Test
             s.OverrideTarget(new Dictionary<string, string>(){ {"folder", "{$APP_FOLDER_PATH}/../test/samples/private/mineseeker/Student_Name_6"} });
             s.Start();
 
-            Assert.AreEqual("Running script DAM - M03 (UF1): Mineseeker (v1.1.0.1):\r\nRunning on single mode:\r\n   Question 1 [0.5 points] - Checking main java file:\r\n      Looking for .java... OK\r\n      Loading the file Main.java... OK\r\n      Compiling the file Main.java... OK\r\n\r\n   Question 2 [1 point] - Main menu:\r\n      Checking the caption... ERROR:\n         -Expected -> %Escull una opció:%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n      Checking the first option... ERROR:\n         -Expected -> %1. Emplenar el taulell%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n      Checking the second option... ERROR:\n         -Expected -> %2. Mostrar el taulell%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n      Checking the third option... ERROR:\n         -Expected -> %3. Sortir%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   Question 3 [1 point] - Input data:\r\n      Checking the input data giving no data... OK\r\n      Checking the input data giving rows... OK\r\n      Checking the input data giving rows and columns... OK\r\n      Checking the input question for the rows... ERROR:\n         -Expected -> %De quantes files vols el taulell?%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nfiles al taulell?\ncolumnes al taulell?\nmines al taulell?\n1 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n      Checking the input question for the columns... ERROR:\n         -Expected -> %De quantes columnes vols el taulell?%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nfiles al taulell?\ncolumnes al taulell?\nmines al taulell?\n1 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n      Checking the input question for the mines... ERROR:\n         -Expected -> %Quantes mines vols al taulell?%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nfiles al taulell?\ncolumnes al taulell?\nmines al taulell?\n1 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   Question 4 [0.5 points] - Data verification:\r\n      Checking that more mines that cells is not allowed... ERROR:\n         -Expected -> %NO ES PODEN POSAR MÉS MINES QUE CASELLES%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nfiles al taulell?\ncolumnes al taulell?\nmines al taulell?\nNO ES PODEN POSAR MÉS MINES\nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   Question 5 [1 point] - Board internal creation:\r\n      Checking the internal board creation... OK\r\n      Checking the internal board sizing... OK\r\n\r\n   Question 6 [1 point] - Board mines filling:\r\n      Checking the mine cell filling... OK\r\n\r\n   Question 7 [1 point] - Random mines positioning:\r\n      Checking that random location is being used... OK\r\n\r\n   Question 8 [0.5 points] - Return to the menu after filling the board:\r\n      Checking the return to the main menu... ERROR:\n         -Expected -> REGEX >0 \"Quantes mines vols al taulell\\?\\s*(?:\\n|\\r\\n)*Escull una opció:\"; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nfiles al taulell?\ncolumnes al taulell?\nmines al taulell?\n1 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   Question 9 [1 point] - Displaying board's empty cells:\r\n      Checking the display for a 1x1 board with no mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking the display for a 5x3 board with no mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking the display for a 3x5 board with no mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking the display for a 5x5 board with no mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n\r\n   Question 10 [1 point] - Displaying board's mine cells:\r\n      Checking the display for a 1x1 board full of mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking the display for a 5x3 board full of mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking the display for a 3x5 board full of mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking the display for a 10x10 board full of mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking that mines are changing its position... ERROR:\n         -Expected -> !=Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nfiles al taulell?\ncolumnes al taulell?\nmines al taulell?\n1 0 0 0 0 0 0 0 0 0 \n0 1 0 0 0 0 0 0 0 0 \n0 0 1 0 0 0 0 0 0 0 \n0 0 0 1 0 0 0 0 0 0 \n0 0 0 0 1 0 0 0 0 0 \n0 0 0 0 0 1 0 0 0 0 \n0 0 0 0 0 0 1 0 0 0 \n0 0 0 0 0 0 0 1 0 0 \n0 0 0 0 0 0 0 0 1 0 \n0 0 0 0 0 0 0 0 0 1 \nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\n[X] [-] [-] [-] [-] [-] [-] [-] [-] [-] \n[-] [X] [-] [-] [-] [-] [-] [-] [-] [-] \n[-] [-] [X] [-] [-] [-] [-] [-] [-] [-] \n[-] [-] [-] [X] [-] [-] [-] [-] [-] [-] \n[-] [-] [-] [-] [X] [-] [-] [-] [-] [-] \n[-] [-] [-] [-] [-] [X] [-] [-] [-] [-] \n[-] [-] [-] [-] [-] [-] [X] [-] [-] [-] \n[-] [-] [-] [-] [-] [-] [-] [X] [-] [-] \n[-] [-] [-] [-] [-] [-] [-] [-] [X] [-] \n[-] [-] [-] [-] [-] [-] [-] [-] [-] [X] \nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nfiles al taulell?\ncolumnes al taulell?\nmines al taulell?\n1 0 0 0 0 0 0 0 0 0 \n0 1 0 0 0 0 0 0 0 0 \n0 0 1 0 0 0 0 0 0 0 \n0 0 0 1 0 0 0 0 0 0 \n0 0 0 0 1 0 0 0 0 0 \n0 0 0 0 0 1 0 0 0 0 \n0 0 0 0 0 0 1 0 0 0 \n0 0 0 0 0 0 0 1 0 0 \n0 0 0 0 0 0 0 0 1 0 \n0 0 0 0 0 0 0 0 0 1 \nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\n[X] [-] [-] [-] [-] [-] [-] [-] [-] [-] \n[-] [X] [-] [-] [-] [-] [-] [-] [-] [-] \n[-] [-] [X] [-] [-] [-] [-] [-] [-] [-] \n[-] [-] [-] [X] [-] [-] [-] [-] [-] [-] \n[-] [-] [-] [-] [X] [-] [-] [-] [-] [-] \n[-] [-] [-] [-] [-] [X] [-] [-] [-] [-] \n[-] [-] [-] [-] [-] [-] [X] [-] [-] [-] \n[-] [-] [-] [-] [-] [-] [-] [X] [-] [-] \n[-] [-] [-] [-] [-] [-] [-] [-] [X] [-] \n[-] [-] [-] [-] [-] [-] [-] [-] [-] [X] \nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   Question 11 [1 point] - Displaying empty board message:\r\n      Checking the empty message... ERROR:\n         -Expected -> REGEX >0 \"NO S(?:'|’)HA CREAT CAP TAULELL\"; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nNO S’HA CREAT\nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   Question 12 [0.5 points] - Return to the menu after displaying the board:\r\n      Checking the menu display... ERROR:\n         -Expected -> REGEX >0 \"NO S(?:'|’)HA CREAT CAP TAULELL\\s*(?:\\n|\\r\\n)*Escull una opció:\"; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nNO S’HA CREAT\nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   TOTAL SCORE: 3.5 / 10", s.Output.ToString());
+            var expected = "Running script DAM - M03 (UF1): Mineseeker (v1.1.0.1):\r\nRunning on single mode:\r\n   Question 1 [0.5 points] - Checking main java file:\r\n      Looking for .java... OK\r\n      Loading the file Main.java... OK\r\n      Compiling the file Main.java... OK\r\n\r\n   Question 2 [1 point] - Main menu:\r\n      Checking the caption... ERROR:\n         -Expected -> %Escull una opció:%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n      Checking the first option... ERROR:\n         -Expected -> %1. Emplenar el taulell%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n      Checking the second option... ERROR:\n         -Expected -> %2. Mostrar el taulell%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n      Checking the third option... ERROR:\n         -Expected -> %3. Sortir%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   Question 3 [1 point] - Input data:\r\n      Checking the input data giving no data... OK\r\n      Checking the input data giving rows... OK\r\n      Checking the input data giving rows and columns... OK\r\n      Checking the input question for the rows... ERROR:\n         -Expected -> %De quantes files vols el taulell?%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nfiles al taulell?\ncolumnes al taulell?\nmines al taulell?\n1 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n      Checking the input question for the columns... ERROR:\n         -Expected -> %De quantes columnes vols el taulell?%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nfiles al taulell?\ncolumnes al taulell?\nmines al taulell?\n1 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n      Checking the input question for the mines... ERROR:\n         -Expected -> %Quantes mines vols al taulell?%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nfiles al taulell?\ncolumnes al taulell?\nmines al taulell?\n1 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   Question 4 [0.5 points] - Data verification:\r\n      Checking that more mines that cells is not allowed... ERROR:\n         -Expected -> %NO ES PODEN POSAR MÉS MINES QUE CASELLES%; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nfiles al taulell?\ncolumnes al taulell?\nmines al taulell?\nNO ES PODEN POSAR MÉS MINES\nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   Question 5 [1 point] - Board internal creation:\r\n      Checking the internal board creation... OK\r\n      Checking the internal board sizing... OK\r\n\r\n   Question 6 [1 point] - Board mines filling:\r\n      Checking the mine cell filling... OK\r\n\r\n   Question 7 [1 point] - Random mines positioning:\r\n      Checking that random location is being used... OK\r\n\r\n   Question 8 [0.5 points] - Return to the menu after filling the board:\r\n      Checking the return to the main menu... ERROR:\n         -Expected -> REGEX >0 \"Quantes mines vols al taulell\\?\\s*(?:\\n|\\r\\n)*Escull una opció:\"; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nfiles al taulell?\ncolumnes al taulell?\nmines al taulell?\n1 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \n0 0 0 0 \nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   Question 9 [1 point] - Displaying board's empty cells:\r\n      Checking the display for a 1x1 board with no mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking the display for a 5x3 board with no mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking the display for a 3x5 board with no mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking the display for a 5x5 board with no mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n\r\n   Question 10 [1 point] - Displaying board's mine cells:\r\n      Checking the display for a 1x1 board full of mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking the display for a 5x3 board full of mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking the display for a 3x5 board full of mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking the display for a 10x10 board full of mines... ERROR:\n         -Expected -> 1; Found -> 0\r\n      Checking that mines are changing its position... OK\r\n\r\n   Question 11 [1 point] - Displaying empty board message:\r\n      Checking the empty message... ERROR:\n         -Expected -> REGEX >0 \"NO S(?:'|’)HA CREAT CAP TAULELL\"; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nNO S’HA CREAT\nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   Question 12 [0.5 points] - Return to the menu after displaying the board:\r\n      Checking the menu display... ERROR:\n         -Expected -> REGEX >0 \"NO S(?:'|’)HA CREAT CAP TAULELL\\s*(?:\\n|\\r\\n)*Escull una opció:\"; Found -> Escull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\n\nNO S’HA CREAT\nEscull la opció: \n   1) Emplenar el taulell\n   2) Mostrar el taulell\n   3) Sortir\r\n\r\n   TOTAL SCORE: 3.5 / 10";
+            Assert.AreEqual(expected, s.Output.ToString());
         }
